requirements.txt
mypy_boto3_s3
numpy
pandas
psycopg2
toposort
